ast: Move annotation json code to own file
matches what we have for term, policy.
